backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0
pyotp==2.9.0
qrcode==7.4.2

# OAuth
authlib==1.3.0
httpx==0.26.0

# Payment Processing
stripe==7.11.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Email
fastapi-mail==1.4.1
jinja2==3.1.3

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.40.0

# File Storage
boto3==1.34.28
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2024.1
pydantic-extra-types==2.4.1
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.5.1

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
